+++ src/main/java/net/sf/jabref/gui/maintable/MainTableSelectionListener.java
@@ -105,24 +105,14 @@ public class MainTableSelectionListener implements ListEventListener<BibEntry>, 
         if (!enabled) {
             return;
         }
+
         EventList<BibEntry> selected = e.getSourceList();
-        BibEntry newSelected = null;
-        while (e.next()) {
-            if (e.getType() == ListEvent.INSERT) {
-                if (newSelected == null) {
-                    if (e.getIndex() < selected.size()) {
-                        newSelected = selected.get(e.getIndex());
-                    }
-                } else {
-                    return; // More than one new selected. Do nothing.
-                }
-            }
+        if (selected.isEmpty()){
+            return;
         }
 
+        final BibEntry newSelected = selected.get(0);
         if (newSelected != null) {
-
-            // Ok, we have a single new entry that has been selected. Now decide what to do with it:
-            final BibEntry toShow = newSelected;
             final BasePanelMode mode = panel.getMode(); // What is the panel already showing?
             if ((mode == BasePanelMode.WILL_SHOW_EDITOR) || (mode == BasePanelMode.SHOWING_EDITOR)) {
                 // An entry is currently being edited.
@@ -132,7 +122,7 @@ public class MainTableSelectionListener implements ListEventListener<BibEntry>, 
                     visName = oldEditor.getVisiblePanelName();
                 }
                 // Get an old or new editor for the entry to edit:
-                EntryEditor newEditor = panel.getEntryEditor(toShow);
+                EntryEditor newEditor = panel.getEntryEditor(newSelected);
 
                 if (oldEditor != null) {
                     oldEditor.setMovingToDifferentEntry();
@@ -150,7 +140,7 @@ public class MainTableSelectionListener implements ListEventListener<BibEntry>, 
             } else {
                 // Either nothing or a preview was shown. Update the preview.
                 if (previewActive) {
-                    updatePreview(toShow, false);
+                    updatePreview(newSelected, false);
                 }
             }
         }